refactor(v2): unit_type/room_type schema 全栈 int 统一 + HIPAA permission 修复
- UnitType/UnitProperty/RoomType 三个 enum 字段 DB SMALLINT ↔ Go int ↔ JSON int 三层对齐
- 新增 owl-common/card.UnitTypePrivate/Share/Public + UnitPropertyHome/Facility +
  RoomTypeDefault/Bathroom/Kitchen 常量；RoomState.Kind string → RoomType int
- 删 BedState.RiskLevel：OOB 单源不构成 risk，作为输入证据喂 sensor fall 链路
  （Still/Bedside/Lost），不再染卡片色
- 删 ipam.UnitAttrs.LayoutType（BE 无业务逻辑读，FE meta 自管）+ IsPublic/IsShared
  （由 unit_type 派生不单存）；UnitAttrs.RoomType string → int
- 删 dead structs：domain.CardOverviewItem + 整个 repository/repository.go
  (Unit/Room/Bed/Device/DeviceStore 老 struct 无 instantiation)
- HIPAA fix: alarm_event_service permission check 修复 silent bug——
  原 unit_type::text vs "home"/"facility" 永远不命中导致权限完全 fallthrough；
  改读 unit_property int 比对 commoncard.UnitPropertyHome/Facility 常量；
  同步删 v2 不存在的 tenant_id/is_public/is_shared_unit 列引用
- domain.Unit / UnitFilters / RoomWithAvailability / spatial_v2 DTO / stub_handlers
  / unit_handler 全部 int8/string → int；helper getInt8 → getIntVal
- spatial_v2 ListRoomsByBranch SQL 去掉 CASE 'Home'/'Facility'/'Private/Share/Public'
  stringify，FE breaking：JSON 字段 unit_type 由字符串变 int + 新增 unit_property
- monitor_handler / card_allowed_provider 改用 UnitProperty 常量代替 unit_type 字符串

+++ wisefido-data/internal/domain/unit.go
@@ -19,11 +19,11 @@ type Unit struct {
 	BuildingName sql.NullString `db:"building_name"`  // nullable, 通过 JOIN buildings 表获取（不存储在 units 表）
 	Floor        sql.NullString `db:"floor"`          // nullable, default '1F' (由 Service 层控制)
 	LayoutConfig sql.NullString `db:"layout_config"`  // nullable, JSONB
-	// v2 双维度（2026-05-09 重设计）：
+	// v2 双维度（与 owl-common/card.UnitProperty* / UnitType* 一致）：
 	//   UnitProperty 0=Home, 1=Facility (default)
-	//   UnitType     0=unknown, 1=single (Private), 2=share (default), 3=public
-	//   约束：Home → UnitType=0；Facility → UnitType ∈ {1,2,3}
-	UnitProperty int8   `db:"unit_property"` // 0=Home, 1=Facility
-	UnitType     int8   `db:"unit_type"`     // 0/1/2/3 enum
+	//   UnitType     1=Private, 2=Share (default for Facility), 3=Public
+	//   约束：Home → UnitType ∈ {1,2}；Facility → UnitType ∈ {1,2,3}
+	UnitProperty int    `db:"unit_property"` // 0=Home, 1=Facility
+	UnitType     int    `db:"unit_type"`     // 1=Private, 2=Share, 3=Public
 	Timezone     string `db:"timezone"`      // NOT NULL
 }
